requirements.txt
pandas
numpy
requests
python-dotenv
statsmodels
matplotlib
scikit-learn